websockets/source/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.3
chardet==4.0.0
idna==3.2
multidict==5.1.0
python-engineio==4.2.1
python-socketio==5.4.0
typing-extensions==3.10.0.2
websockets==10.0
yarl==1.6.3
